Fix parseArguments, refactor, implement swift theme files generation
