+++ OriginateTheme/Scripts/ot_generator_templates.py
@@ -1,0 +1,18 @@
+###########################
+##### Swift Templates #####
+###########################
+
+def swiftTemplate():
+    """
+        Template for an OriginateTheme swift file.
+    """
+    return """//
+//  $OriginateThemeStructName.swift
+//  OriginateTheme
+//
+//  Copyright (c) 2017 Originate. All rights reserved.
+//
+
+import UIKit
+
+"""
